Working solution for Day 1 in SQL
Part 2 is kind of a mess. For some reason trying to use
GROUP BY / HAVING for it results in the query hanging. So instead
I've opted to use a UNIQUE constraint on a table and just doing row
inserts, Postgresql will then tell me when it hits the duplicate row.